[nrf noup] cmake: support CONFIG_MCUBOOT_BUILD_S1_VARIANT
Support the nRF Connect SDK-specfic Kconfig option
CONFIG_MCUBOOT_BUILD_S1_VARIANT. Note that this option is defined in
the fw-nrfconnect-nrf repository in the NCS. The changes required to
support it are being applied to the zephyr repository out of necessity
rather than choice: specifically, it is non-trivial to add executable
targets outside the zephyr/CMakeLists.txt. The problem is that most of
the variables required for and compiling/linking are only available
here.

Register in the parent image that the mcuboot child image will have
s1_image.hex as a byproduct, this allows the parent image to know
where the hex file comes from and create custom commands that sign it.
Ensure that the 's1_image_hex' target is built when the ALL target is
built so that multi-image builds work correctly.

Signed-off-by: Håkon Øye Amundsen <haakon.amundsen@nordicsemi.no>
Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>
Signed-off-by: Robert Lubos <robert.lubos@nordicsemi.no>
Signed-off-by: Dominik Ermel <dominik.ermel@nordicsemi.no>
Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>