[nrf toup] bluetooth: Selecting ENTROPY_GENERATOR for crypto
Bluetooth sample with crypto, in NRF, requires sys_rand32_get that
used to be linked with Tinycrypt. The selection, within Kconfig of
Tinycrypt, that has been enabling compilation of the symbol has
been removed with commit zephyrpoject-rtos/zephyr:e07227ff
as a result preventing NRF samples to link.

This commit moves the selection to BT_CTLR_CRYPTO, which allows
sample to link again.

Signed-off-by: Dominik Ermel <dominik.ermel@nordicsemi.no>