[nrf temphack] west: REVERTME: Change variable used to find app dir
APPLICATION_SOURCE_DIR was formerly not in the cache due to
multi-image requiring cache variables to be image-prefixed and due to
a desire to not image-prefix this variable.

However, west expected this variable in the ache. To resolve this, we
had west use a different variable that points to the same thing. This
commit is now no longer needed and is safe to revert after the next
history rewrite.

Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>
Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>