[nrf noup] partition manager: Remove dead code
Remove the de-reference of mcuboot_logical_target. It is not set at
that point in the execution and always evaluates to the empty string,
making it do nothing.

Also, it is not correct that preprocessing the S1 variants build
script should depend on linking the S0 elf file. So the dependency is
unnecessary.

Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>